Move tests into their own directory
