+++ com.unity.multiplayer.mlapi/Tests/Runtime/Metrics/Messaging/NetworkMetricsMessagingDispatchTests.cs
@@ -111,29 +111,6 @@ namespace MLAPI.RuntimeTests.Metrics.Messaging
 
             yield return WaitForMetricsDispatch();
         }
-
-        [UnityTest]
-        public IEnumerator NetworkMetrics_WhenNetworkObjectSpawnSentToClient_TracksObjectSpawnSentMetric()
-        {
-            const string objectName = "TestNetworkObjectToSpawn";
-            const ulong clientId = 100;
-            var gameObjectId = NetworkManagerHelper.AddGameNetworkObject(objectName);
-            NetworkManagerHelper.InstantiatedNetworkObjects.TryGetValue(gameObjectId, out var networkObj);
-            NetworkManagerHelper.AddConnectedClient(clientId);
-
-            m_NetworkMetrics.Dispatcher.RegisterObserver(new TestObserver(collection =>
-            {
-                var objectSpawnedMetric = AssertSingleMetricEventOfType<ObjectSpawnedEvent>(collection, MetricNames.ObjectSpawned);
-                Assert.AreEqual(1, objectSpawnedMetric.Values.Count);
-                Assert.AreEqual(clientId, objectSpawnedMetric.Values.Select(x => x.Connection.Id).First());
-                Assert.AreEqual(objectName, objectSpawnedMetric.Values.Select(x => x.NetworkId.Name));
-                m_NetworkManager.SpawnManager.DespawnObject(networkObj);
-            }));
-            networkObj.Observers.Add(clientId);
-            networkObj.Spawn();
-
-            yield return WaitForMetricsDispatch();
-        }
     }
 #endif
 }
