+++ com.unity.multiplayer.mlapi/Tests/Runtime/Metrics/NetworkObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9ffac4e9e439f0a4d81c0cfc438851d4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
