+++ com.unity.multiplayer.mlapi/Tests/Runtime/Metrics/NetworkObjects/NetworkObjectMetricsDispatchTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 873332ad5dde52e4e9fdd73858e77b75
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
